Test /dist/$dist/$version/$path.html.
Make sure it properly handles 404s. In passing, tweak the distribution
template to properly highlight the distribution link, and tweak the document
template to make the distribution link URL a path rather than a full URL.
Also, change the arrow separator in the breadcrumb to something a little less
huge.