Add /users/, search for users by first char of nickname.
This is mostly so that search bots can find every user on the site, and from
each user, every distribution and extension (and tag). Implemented using
wildcard searches, so pretty simple.

Other stuff:

* Users link is in top nav, so move the "Release It" link to the footer. Now
  all info links are in the nav, and action links (releasing, sending
  feedback, etc.) are in the footer.
* Added link to Younicycle, new benefactor!
* Abstracted the display of search results into their own template, so that
  both search and /users can use it.
* Tweaked display on /backers/: Use h2 instead of h3 for subheads, but change
  its color to Yellow so it's easier to distinguish from all the backer links.

I think that we're now feature-complete for a 1.0. Will deploy for a limited
beta on Monday and continue writing tests.