Fix TOC display on Firefox.
Downside is that the background color on hover now starts at the beginning of
the text. Before, on Safari, it would show about 8px to the left of the text,
too. I'd like to make it so that it goes all the way across the sidebar, under
the bullet, too. Will have to figure that out later.