Simplify nav and footer links.
Nav links are now only about content, while the footer is external links and
metadata. The donor and identity links are gone, but still referenced from the
about and FAQ pages. Also added a "Colophon" to the about page to credit the
design, since thoese links have also been removed from the footer. The site
version now only appears in a "generator" meta tag in the header.