Parse and format Multimarkdown at release time
Instead of shipping the .mmd files, ship .html files generated by the
`multimarkdown` CLI. Requires a bit of tweaking of the markdown files,
but allows the use of fenced code blocks and eliminates a build-time
dependency.

Also tweak the CSS to open up some space in definition lists, though not
in the meta section of a distribution page. Looks nicer, and should be
easier to read.